Added check whether the autorun feature is supported to the Disable() method of the AutorunManager class.

+++ src/lib/autorunmanager/autorunmanager.cpp
@@ -90,6 +90,6 @@ void AutorunManager::Enable() const
 
 void AutorunManager::Disable() const
 {
-    if (IsEnabled())
+    if (AutorunSupported && IsEnabled())
         QFile::remove(AutorunFileName);
 }
